measures(phase-8): absorb scripts/measure-relative-imports + measure-relative-paths into health/{coupling,shape}/ vitest tests
Per tasks.md Phase 8 sub-steps 8.1–8.4:

- Create packages/measures/src/health/coupling/relative-import-depth.test.ts
  absorbing scripts/measure-relative-imports.mjs logic; budget carried
  verbatim from measure-relative-imports.report.json.
- Create packages/measures/src/health/shape/relative-path-depth.test.ts
  absorbing scripts/measure-relative-paths.mjs logic; budget carried
  verbatim from measure-relative-paths.report.json.
- Verify/update external/tier_1/relative-{import,path}-depth.ts
  CheckDef stubs as documented in 8.3.
- Delete scripts/measure-relative-imports.mjs,
  measure-relative-paths.mjs, and their .report.json companions.

Phase 8 uses the Phase-7-snapshot baseline (brain 180b23e). Phase 3 +
Phase 4 + Phase 7 ratification sidecars are historical. Fresh
Phase-7-epoch sidecar set starts empty; no carry-forward entries
needed for this phase.

After this commit, scripts/measure-*.mjs is empty (only check-* remaining
have been absorbed in earlier phases; this absorbs the structural-measure
family).

Verifier strategy: strict gate against Phase-7-snapshot baseline. Schemas
of relative-import-depth.json + relative-path-depth.json preserved
verbatim. No structural shifts expected (new tests are small .test.ts
files; isProductionSource excludes them from production metrics; not
large enough to cascade top-N largestFiles per Mia's discipline note).
