+++ packages/measures/src/external/tier_1/relative-import-depth.ts
@@ -1,10 +1,16 @@
-import {type CheckDef} from '../../_types.ts'
+import {type CheckDef, vitestJsonArgs} from '../../_types.ts'
 
 export const check: CheckDef = {
     id: 'relative-import-depth',
     name: 'Relative Import Depth',
     category: 'Static',
-    display: 'node scripts/measure-relative-imports.mjs --enforce',
-    args: () => ['node', 'scripts/measure-relative-imports.mjs', '--enforce'],
-    parser: 'none',
+    display: 'vitest run packages/measures/src/health/coupling/relative-import-depth.test.ts',
+    args: (jsonOut) => [
+        'npx',
+        'vitest',
+        'run',
+        'packages/measures/src/health/coupling/relative-import-depth.test.ts',
+        ...vitestJsonArgs(jsonOut),
+    ],
+    parser: 'vitest',
 }
