+++ packages/measures/src/external/tier_1/relative-path-depth.ts
@@ -1,10 +1,16 @@
-import {type CheckDef} from '../../_types.ts'
+import {type CheckDef, vitestJsonArgs} from '../../_types.ts'
 
 export const check: CheckDef = {
     id: 'relative-path-depth',
     name: 'Relative Path Depth',
     category: 'Static',
-    display: 'node scripts/measure-relative-paths.mjs --enforce',
-    args: () => ['node', 'scripts/measure-relative-paths.mjs', '--enforce'],
-    parser: 'none',
+    display: 'vitest run packages/measures/src/health/shape/relative-path-depth.test.ts',
+    args: (jsonOut) => [
+        'npx',
+        'vitest',
+        'run',
+        'packages/measures/src/health/shape/relative-path-depth.test.ts',
+        ...vitestJsonArgs(jsonOut),
+    ],
+    parser: 'vitest',
 }
