--- scripts/measure-relative-imports.mjs
@@ -1,402 +1,0 @@
-import {existsSync, readFileSync, statSync, writeFileSync} from 'node:fs'
-import {readdir, stat} from 'node:fs/promises'
-import {dirname, join, relative, resolve, sep} from 'node:path'
-import {fileURLToPath} from 'node:url'
-import * as ts from 'typescript'
-
-const REPO_ROOT = resolve(dirname(fileURLToPath(import.meta.url)), '..')
-const REPORT_PATH = join(REPO_ROOT, 'scripts', 'measure-relative-imports.report.json')
-const ENFORCE = process.argv.includes('--enforce')
-
-const REQUESTED_SCOPE_ROOTS = [
-  'webapp/src',
-  'packages/systems/*/src',
-  'packages/libraries/*/src',
-  'packages/systems/voicetree-mcp/bin',
-]
-
-const EXCLUDED_DIR_NAMES = new Set(['node_modules', 'dist', 'build', '__tests__', 'integration-tests'])
-const RESOLUTION_EXTENSIONS = ['', '.ts', '.tsx', '.d.ts', '.js', '.jsx', '.json']
-
-const sourceRoots = await discoverSourceRoots()
-const files = await listProductionSourceFiles(sourceRoots)
-if (files.length === 0) {
-  throw new Error('measure-relative-imports found 0 production source files; check source globs')
-}
-
-const packageRootCache = new Map()
-const packageNameCache = new Map()
-const records = collectRelativeImportRecords(files)
-const summary = summarizeRecords(records, files.length)
-
-writeFileSync(REPORT_PATH, `${JSON.stringify(summary, null, 2)}\n`)
-printReport(summary)
-if (ENFORCE && summary.totals.bannedRelativeImports > 0) {
-  console.error(`Relative import depth gate failed: ${summary.totals.bannedRelativeImports} banned relative import(s) found`)
-  process.exit(1)
-}
-
-async function discoverSourceRoots() {
-  const packageSrcRoots = await Promise.all([
-    discoverPackageSrcRoots(join(REPO_ROOT, 'packages', 'systems')),
-    discoverPackageSrcRoots(join(REPO_ROOT, 'packages', 'libraries')),
-  ])
-
-  return [
-    join(REPO_ROOT, 'webapp', 'src'),
-    ...packageSrcRoots.flat(),
-    join(REPO_ROOT, 'packages', 'systems', 'voicetree-mcp', 'bin'),
-  ]
-}
-
-async function discoverPackageSrcRoots(layerRoot) {
-  if (!(await pathExists(layerRoot))) return []
-  const entries = await readdir(layerRoot, {withFileTypes: true})
-  return entries
-    .filter(entry => entry.isDirectory())
-    .map(entry => join(layerRoot, entry.name, 'src'))
-}
-
-async function listProductionSourceFiles(roots) {
-  const nested = await Promise.all(roots.map(async root => {
-    const absoluteRoot = root
-    if (!(await pathExists(absoluteRoot))) return []
-    return listProductionSourcesUnder(absoluteRoot)
-  }))
-
-  return [...new Set(nested.flat())].sort((a, b) => relative(REPO_ROOT, a).localeCompare(relative(REPO_ROOT, b)))
-}
-
-async function listProductionSourcesUnder(root) {
-  const entries = await readdir(root, {withFileTypes: true})
-  const nested = await Promise.all(entries.map(async entry => {
-    const absolutePath = join(root, entry.name)
-    if (entry.isDirectory()) {
-      if (EXCLUDED_DIR_NAMES.has(entry.name)) return []
-      return listProductionSourcesUnder(absolutePath)
-    }
-    if (!entry.isFile() || !isProductionTypeScriptSource(absolutePath)) return []
-    return [absolutePath]
-  }))
-  return nested.flat()
-}
-
-function isProductionTypeScriptSource(path) {
-  return /\.(ts|tsx)$/.test(path)
-    && !/\.d\.ts$/.test(path)
-    && !/\.(test|spec)\.(ts|tsx)$/.test(path)
-    && !/\.config\.ts$/.test(path)
-    && !containsExcludedSegment(path)
-}
-
-function containsExcludedSegment(path) {
-  const segments = relative(REPO_ROOT, path).split(sep)
-  return segments.some(segment => EXCLUDED_DIR_NAMES.has(segment))
-}
-
-async function pathExists(path) {
-  try {
-    await stat(path)
-    return true
-  } catch {
-    return false
-  }
-}
-
-function collectRelativeImportRecords(sourceFiles) {
-  return sourceFiles.flatMap(file => {
-    const text = readFileSync(file, 'utf8')
-    const sourceFile = ts.createSourceFile(file, text, ts.ScriptTarget.Latest, true, file.endsWith('.tsx') ? ts.ScriptKind.TSX : ts.ScriptKind.TS)
-    const packageRoot = findPackageRoot(file)
-    const packageName = packageNameForRoot(packageRoot)
-
-    return extractModuleSpecifiers(sourceFile)
-      .filter(importSite => importSite.specifier.startsWith('.'))
-      .map(importSite => classifyRelativeImport(file, packageRoot, packageName, importSite, sourceFile))
-  })
-}
-
-function extractModuleSpecifiers(sourceFile) {
-  const importSites = []
-
-  function visit(node) {
-    if (ts.isImportDeclaration(node) && ts.isStringLiteralLike(node.moduleSpecifier)) {
-      importSites.push(importSite('static', node.moduleSpecifier, sourceFile))
-    } else if (ts.isExportDeclaration(node) && node.moduleSpecifier && ts.isStringLiteralLike(node.moduleSpecifier)) {
-      importSites.push(importSite('re-export', node.moduleSpecifier, sourceFile))
-    } else if (ts.isCallExpression(node) && node.expression.kind === ts.SyntaxKind.ImportKeyword) {
-      const [firstArg] = node.arguments
-      if (firstArg && ts.isStringLiteralLike(firstArg)) importSites.push(importSite('dynamic', firstArg, sourceFile))
-    } else if (ts.isImportTypeNode(node) && ts.isLiteralTypeNode(node.argument) && ts.isStringLiteralLike(node.argument.literal)) {
-      importSites.push(importSite('import-type', node.argument.literal, sourceFile))
-    } else if (ts.isImportEqualsDeclaration(node) && ts.isExternalModuleReference(node.moduleReference)) {
-      const expression = node.moduleReference.expression
-      if (ts.isStringLiteralLike(expression)) importSites.push(importSite('import-equals', expression, sourceFile))
-    }
-
-    ts.forEachChild(node, visit)
-  }
-
-  visit(sourceFile)
-  return importSites
-}
-
-function importSite(kind, literal, sourceFile) {
-  const {line, character} = sourceFile.getLineAndCharacterOfPosition(literal.getStart(sourceFile))
-  return {
-    kind,
-    specifier: literal.text,
-    line: line + 1,
-    column: character + 1,
-  }
-}
-
-function classifyRelativeImport(importerPath, importerPackageRoot, importerPackageName, importSite, sourceFile) {
-  const resolvedPath = resolveRelativeSpecifier(importerPath, importSite.specifier)
-  const targetPackageRoot = findPackageRoot(resolvedPath)
-  const targetPackageName = packageNameForRoot(targetPackageRoot)
-  const isCrossPackage = targetPackageRoot !== importerPackageRoot
-
-  return {
-    importer: normalizePath(relative(REPO_ROOT, importerPath)),
-    line: importSite.line,
-    column: importSite.column,
-    kind: importSite.kind,
-    specifier: importSite.specifier,
-    depth: relativeDepthBucket(importSite.specifier),
-    classification: isCrossPackage ? 'relative-cross-package' : 'relative-same-package',
-    importerPackage: importerPackageName,
-    importerPackageRoot: normalizePath(relative(REPO_ROOT, importerPackageRoot)),
-    resolvedPath: normalizePath(relative(REPO_ROOT, resolvedPath)),
-    targetPackage: targetPackageName,
-    targetPackageRoot: normalizePath(relative(REPO_ROOT, targetPackageRoot)),
-    statement: statementTextAt(sourceFile, importSite.line),
-  }
-}
-
-function resolveRelativeSpecifier(importerPath, specifier) {
-  const basePath = resolve(dirname(importerPath), specifier)
-  const candidates = [
-    ...RESOLUTION_EXTENSIONS.map(extension => `${basePath}${extension}`),
-    ...RESOLUTION_EXTENSIONS.filter(Boolean).map(extension => join(basePath, `index${extension}`)),
-  ]
-  return candidates.find(candidate => existsSync(candidate)) ?? basePath
-}
-
-function findPackageRoot(path) {
-  const startingDirectory = directoryForPackageLookup(path)
-  const cached = packageRootCache.get(startingDirectory)
-  if (cached) return cached
-
-  let current = startingDirectory
-  while (true) {
-    if (existsSync(join(current, 'package.json'))) {
-      packageRootCache.set(startingDirectory, current)
-      return current
-    }
-    const parent = dirname(current)
-    if (parent === current) {
-      packageRootCache.set(startingDirectory, REPO_ROOT)
-      return REPO_ROOT
-    }
-    current = parent
-  }
-}
-
-function directoryForPackageLookup(path) {
-  if (existsSync(path)) {
-    try {
-      return statSync(path).isDirectory() ? path : dirname(path)
-    } catch {
-      return dirname(path)
-    }
-  }
-  return /\.[a-z0-9]+$/i.test(path) ? dirname(path) : path
-}
-
-function packageNameForRoot(packageRoot) {
-  const cached = packageNameCache.get(packageRoot)
-  if (cached) return cached
-
-  const packageJsonPath = join(packageRoot, 'package.json')
-  const packageJson = JSON.parse(readFileSync(packageJsonPath, 'utf8'))
-  const name = packageJson.name ?? (normalizePath(relative(REPO_ROOT, packageRoot)) || 'repo-root')
-  packageNameCache.set(packageRoot, name)
-  return name
-}
-
-function relativeDepthBucket(specifier) {
-  if (specifier.startsWith('./')) return './x'
-  const matches = specifier.match(/^(\.\.\/)+/)
-  const depth = matches ? matches[0].split('../').length - 1 : 0
-  if (depth <= 1) return '../x'
-  if (depth === 2) return '../../x'
-  return '../../../+x'
-}
-
-function statementTextAt(sourceFile, oneBasedLine) {
-  const lineStarts = sourceFile.getLineStarts()
-  const start = lineStarts[oneBasedLine - 1] ?? 0
-  const end = lineStarts[oneBasedLine] ?? sourceFile.getFullText().length
-  return sourceFile.getFullText().slice(start, end).trim()
-}
-
-function summarizeRecords(allRecords, sourceFileCount) {
-  const samePackageRecords = allRecords.filter(record => record.classification === 'relative-same-package')
-  const crossPackageRecords = allRecords.filter(record => record.classification === 'relative-cross-package')
-  const deepSamePackageRecords = samePackageRecords.filter(record => record.depth === '../../x' || record.depth === '../../../+x')
-  const bannedRecords = [...crossPackageRecords, ...deepSamePackageRecords]
-
-  return {
-    generatedAt: new Date().toISOString(),
-    sourceFileCount,
-    scope: {
-      requestedRoots: REQUESTED_SCOPE_ROOTS,
-      scannedRoots: sourceRoots.map(root => normalizePath(relative(REPO_ROOT, root))),
-      excluded: [
-        '.test.ts',
-        '.spec.ts',
-        '.d.ts',
-        '__tests__',
-        'integration-tests',
-        'node_modules',
-        'dist',
-        'build',
-        '*.config.ts',
-      ],
-    },
-    totals: {
-      relativeImports: allRecords.length,
-      samePackageRelativeImports: samePackageRecords.length,
-      crossPackageRelativeImports: crossPackageRecords.length,
-      deepSamePackageRelativeImports: deepSamePackageRecords.length,
-      bannedRelativeImports: bannedRecords.length,
-    },
-    samePackageByDepth: countBy(samePackageRecords, record => record.depth, ['./x', '../x', '../../x', '../../../+x']),
-    perPackageTotals: summarizePerPackage(allRecords),
-    worstOffenders: summarizeWorstOffenders(allRecords),
-    bannedRelatives: bannedRecords.map(record => ({
-      importer: record.importer,
-      line: record.line,
-      specifier: record.specifier,
-      kind: record.kind,
-      classification: record.classification,
-      depth: record.depth,
-      importerPackage: record.importerPackage,
-      targetPackage: record.targetPackage,
-      resolvedPath: record.resolvedPath,
-      statement: record.statement,
-    })),
-    crossPackageRelatives: crossPackageRecords.map(record => ({
-      importer: record.importer,
-      line: record.line,
-      specifier: record.specifier,
-      kind: record.kind,
-      importerPackage: record.importerPackage,
-      targetPackage: record.targetPackage,
-      resolvedPath: record.resolvedPath,
-      statement: record.statement,
-    })),
-  }
-}
-
-function summarizePerPackage(records) {
-  const packages = new Map()
-  for (const record of records) {
-    const data = packages.get(record.importerPackage) ?? {
-      package: record.importerPackage,
-      packageRoot: record.importerPackageRoot,
-      relativeImports: 0,
-      samePackageRelativeImports: 0,
-      crossPackageRelativeImports: 0,
-      samePackageByDepth: {'./x': 0, '../x': 0, '../../x': 0, '../../../+x': 0},
-    }
-
-    data.relativeImports += 1
-    if (record.classification === 'relative-cross-package') {
-      data.crossPackageRelativeImports += 1
-    } else {
-      data.samePackageRelativeImports += 1
-      data.samePackageByDepth[record.depth] += 1
-    }
-    packages.set(record.importerPackage, data)
-  }
-
-  return [...packages.values()].sort((a, b) =>
-    b.relativeImports - a.relativeImports || a.package.localeCompare(b.package))
-}
-
-function summarizeWorstOffenders(records) {
-  const byFile = new Map()
-  for (const record of records) {
-    const data = byFile.get(record.importer) ?? {
-      file: record.importer,
-      package: record.importerPackage,
-      relativeImports: 0,
-      crossPackageRelativeImports: 0,
-    }
-    data.relativeImports += 1
-    if (record.classification === 'relative-cross-package') data.crossPackageRelativeImports += 1
-    byFile.set(record.importer, data)
-  }
-
-  return [...byFile.values()]
-    .sort((a, b) =>
-      b.crossPackageRelativeImports - a.crossPackageRelativeImports
-      || b.relativeImports - a.relativeImports
-      || a.file.localeCompare(b.file))
-    .slice(0, 20)
-}
-
-function countBy(records, keyFn, orderedKeys) {
-  const counts = Object.fromEntries(orderedKeys.map(key => [key, 0]))
-  for (const record of records) {
-    counts[keyFn(record)] = (counts[keyFn(record)] ?? 0) + 1
-  }
-  return counts
-}
-
-function printReport(summary) {
-  console.log('Relative import measurement report')
-  console.log('')
-  console.log(`Source files scanned: ${summary.sourceFileCount}`)
-  console.log(`Option B cost, all relative imports: ${summary.totals.relativeImports}`)
-  console.log(`Option A cost, cross-package relative imports: ${summary.totals.crossPackageRelativeImports}`)
-  console.log(`Same-package relative imports: ${summary.totals.samePackageRelativeImports}`)
-  console.log(`Banned by depth policy (cross-package or ../../+): ${summary.totals.bannedRelativeImports}`)
-  console.log('')
-  console.log('Same-package relative imports by depth:')
-  for (const [depth, count] of Object.entries(summary.samePackageByDepth)) {
-    console.log(`  ${depth}: ${count}`)
-  }
-  console.log('')
-  console.log('Per-package totals:')
-  for (const pkg of summary.perPackageTotals) {
-    console.log(`  ${pkg.package}: ${pkg.relativeImports} total, ${pkg.crossPackageRelativeImports} cross-package, ${pkg.samePackageRelativeImports} same-package`)
-  }
-  console.log('')
-  console.log('Cross-package relative imports:')
-  if (summary.crossPackageRelatives.length === 0) {
-    console.log('  none')
-  } else {
-    for (const record of summary.crossPackageRelatives) {
-      console.log(`  ${record.importer}:${record.line} ${record.specifier} -> ${record.resolvedPath}`)
-    }
-  }
-  console.log('')
-  console.log('Deep same-package relative imports:')
-  const deepSamePackage = summary.bannedRelatives.filter(record => record.classification === 'relative-same-package')
-  if (deepSamePackage.length === 0) {
-    console.log('  none')
-  } else {
-    for (const record of deepSamePackage) {
-      console.log(`  ${record.importer}:${record.line} ${record.specifier} -> ${record.resolvedPath}`)
-    }
-  }
-  console.log('')
-  console.log(`JSON report written to ${normalizePath(relative(REPO_ROOT, REPORT_PATH))}`)
-}
-
-function normalizePath(path) {
-  return path.split(sep).join('/')
-}
